* daemon.py: no config found shutdown after 1 hr; touchups * deploy.sh: turn on setgid bit for executable * qstat.py: don't convert to category, auto dtypes (maybe temporary)

+++ daemon.py
@@ -14,7 +14,8 @@ from ConfigParser import ConfigParser
 #from datetime import datetime
 from getpass import getuser
 import os
-from os import dup2, fork, getpid, kill, listdir, remove, setgid, setsid, umask
+from os import (chdir, dup2, fork, getpid, kill, listdir, remove, setgid,
+                setsid, umask)
 from os.path import abspath, dirname, getmtime, join, isdir, isfile, ismount
 from random import randint
 from shutil import copy2, copytree, rmtree
@@ -22,7 +23,7 @@ import signal
 from socket import gethostname
 from subprocess import CalledProcessError, check_output, STDOUT
 import sys
-from time import sleep
+from time import sleep, time
 
 if __name__ == '__main__' and __package__ is None:
     os.sys.path.append(dirname(dirname(abspath(__file__))))
@@ -53,7 +54,9 @@ SIGNAL_MAP = {
 }
 """Mapping from signal numbers to their names"""
 
-RESPAWN_DELAY_SEC = 10
+RESPAWN_DELAY_SEC = 60
+NO_CONFIG_SHUTDOWN_SEC = 3600
+"""Shutdown if a config file is not found after this long"""
 
 
 # TODO: get signal handling / respawning working.
@@ -61,7 +64,7 @@ RESPAWN_DELAY_SEC = 10
 #    signame = SIGNAL_MAP[signum]
 #    wstderr('Received signum %d (%s); relaunching daemon.\n'
 #            % (signum, signame))
-#    out = check_output(join(self.configdir, 'deploy.sh'))
+#    out = check_output(join(self.config_dir, 'deploy.sh'))
 #
 #
 #for sigkey in dir(signal):
@@ -74,6 +77,12 @@ RESPAWN_DELAY_SEC = 10
 #        print ("Skipping {}".format(i))
 
 
+def sig_handler(signum, frame):
+    """Handle signals"""
+    wstderr(str(frame))
+    sys.exit(signum)
+
+
 class Daemon(object):
     """
     Queue daemon
@@ -83,18 +92,24 @@ class Daemon(object):
     configfile : string
 
     """
-    def __init__(self, configfile, daemon=True):
+    def __init__(self, configfile, daemon=True, logfile=None):
         self.myusername = getuser()
-        self.config = None
-        self.configfile = expand(configfile)
-        self.configdir = dirname(self.configfile)
-        self.distfile = join(self.configdir, 'dist', 'daemon')
-        self.config_time = 0
-        self.qstat = None
-        self.upgrade_kind = False
+        if logfile is not None:
+            logfile = expand(logfile)
+        self.logfile = logfile
+
+        self.config_parser = None
+        self.config_fpath = expand(configfile)
+        self.config_dir = dirname(self.config_fpath)
+        self.config_file_mtime = 0
         self.configured = False
-        self.reconf()
-        self.queue_stat = {'q': 0, 'r': 0, 'other': 0}
+        self.config_last_seen = 0
+
+        self.upgrade_kind = None
+        self.distfile = join(self.config_dir, 'dist', 'daemon')
+        """File to use when upgrading"""
+
+        self.queue_status = {'q': 0, 'r': 0, 'other': 0}
         self.make_daemon = daemon
         self.is_daemon = False
         self.pid_fpath = expand('~/.pid')
@@ -103,27 +118,37 @@ class Daemon(object):
         self.stderr = None
         self.pid = getpid()
         self.hostname = gethostname()
+        self.qstat = None
+        self.reconfigure()
 
-    def reconf(self):
+    def reconfigure(self):
         """If configfile has been updated, reconfigure accordingly"""
         # First check if that file was touched
         try:
-            config_time = getmtime(self.configfile)
+            config_file_mtime = getmtime(self.config_fpath)
         except OSError, err:
             if self.configured and err.args[0] == 2:
+                dt_sec = time() - self.config_last_seen
+                if dt_sec > NO_CONFIG_SHUTDOWN_SEC:
+                    wstderr('No config file seen in at least %d sec. Shutting'
+                            ' down.' % dt_sec)
+                    sys.exit(1)
                 wstderr('Could not find config "%s", returning to normal'
-                        ' operation\n' % self.configfile)
+                        ' operation\n' % self.config_fpath)
                 return
             raise
+        else:
+            self.config_last_seen = time()
 
-        if config_time <= self.config_time:
+        if config_file_mtime <= self.config_file_mtime:
             return
 
         wstderr('Updated/new config detected; reconfiguring...\n')
-        self.config = ConfigParser()
-        files_read = self.config.read(self.configfile)
+        self.config_parser = ConfigParser()
+        files_read = self.config_parser.read(self.config_fpath)
         if not files_read:
-            wstderr('Config file "%s" could not be read... ' % self.configfile)
+            wstderr('Config file "%s" could not be read... '
+                    % self.config_fpath)
             if self.configured:
                 wstderr('Returning to normal operation using previously-read'
                         ' config.\n')
@@ -131,14 +156,15 @@ class Daemon(object):
                 wstderr('No configuration exists; exiting.\n')
                 sys.exit(1)
 
-        self.users = self.config.get('Users', 'list').split(',')
+        self.users = [u.strip() for u in
+                      self.config_parser.get('Users', 'list').split(',')]
         if self.myusername not in self.users:
             raise ValueError(
                 'User running script "%s" is not in users list %s'
                 % self.myusername, self.users
             )
 
-        self.group = self.config.get('Users', 'group')
+        self.group = self.config_parser.get('Users', 'group')
         self.gid = None
         try:
             self.gid = get_gid(self.group)
@@ -146,21 +172,21 @@ class Daemon(object):
             pass
         self.set_my_umask_gid()
 
-        for key, _ in self.config.items('Directories'):
+        for key, _ in self.config_parser.items('Directories'):
             self.setup_dir(key)
-        self.n_run = int(self.config.get('Queue', 'n_run'))
-        self.n_queue = int(self.config.get('Queue', 'n_queue'))
-        self.sleep = int(self.config.get('Queue', 'sleep'))
+        self.n_run = int(self.config_parser.get('Queue', 'n_run'))
+        self.n_queue = int(self.config_parser.get('Queue', 'n_queue'))
+        self.sleep = int(self.config_parser.get('Queue', 'sleep'))
         self.qstat_cache_dir = expand(
-            self.config.get('Logging', 'qstat_cache_dir')
+            self.config_parser.get('Logging', 'qstat_cache_dir')
         )
         self.mkdir(self.qstat_cache_dir)
         self.qstat = QstatBase(stale_sec=self.sleep - 1,
                                cache_dir=self.qstat_cache_dir,
                                group=self.group)
 
-        if self.config.has_section('Commands'):
-            commands = self.config.items('Commands')
+        if self.config_parser.has_section('Commands'):
+            commands = self.config_parser.items('Commands')
             wstderr('Found [Commands] section, with following commands:\n')
             for command in commands:
                 wstderr('> %s = "%s"\n' % command) # DEBUG
@@ -175,7 +201,7 @@ class Daemon(object):
                         wstderr('Shutdown command issued.\n')
                         sys.exit(0)
 
-        self.config_time = config_time
+        self.config_file_mtime = config_file_mtime
         self.configured = True
 
     # TODO: do the upgrade ourselves here, don't call the shell script!
@@ -228,6 +254,7 @@ class Daemon(object):
             wstderr('Backup dir "%s" is a mount point; refusing to modify.\n'
                     % backupdir)
             wstderr('Cointinuing operation without upgrading!\n')
+            return
         elif isdir(backupdir):
             wstderr('Backup dir "%s" already exists; removing.\n' % backupdir)
             rmtree(backupdir)
@@ -236,15 +263,19 @@ class Daemon(object):
                     % backupdir)
             remove(backupdir)
 
-        # Perform the actual backup of the current files
+        # Backup the current files
         copytree(APPLICATION_DIR, backupdir)
         rename_or_move(self.pid_fpath, backup_pid_fpath)
+        orig_pid_fpath = self.pid_fpath
+        self.pid_fpath = None
         set_path_metadata(backup_pid_fpath, group=self.group, perms=0o660)
 
         wstderr('2. Running `deploy.sh` to get and launch new version...\n')
+        wstderr('-'*79 + '\n')
         try:
-            out = check_output(join(self.configdir, 'deploy.sh'))
+            out = check_output(join(self.config_dir, 'deploy.sh'))
         except CalledProcessError:
+            wstderr('-'*79 + '\n')
             wstderr('> Failed to deploy new software. Reverting code and'
                     ' continuing to run.\n')
 
@@ -256,18 +287,17 @@ class Daemon(object):
             remove_contents(APPLICATION_DIR)
 
             copy_contents(backupdir, APPLICATION_DIR)
-            copy2(backup_pid_fpath, self.pid_fpath)
+            copy2(backup_pid_fpath, orig_pid_fpath)
             rmtree(backupdir)
             remove(backup_pid_fpath)
+            self.pid_fpath = orig_pid_fpath
         else:
+            wstderr('-'*79 + '\n')
             wstderr('> Upgrade appears to have succeeded; output of'
                     ' deploy.sh:\n%s\n' % out)
             wstderr('> Cleaning up and shutting down obsolete daemon...\n')
             rmtree(backupdir)
             remove(backup_pid_fpath)
-            # Don't want to remove PID file since it's being used by another
-            # process now
-            self.pid_fpath = None
             sys.exit(0)
 
     def getpath(self, dir_kind, usr):
@@ -285,12 +315,11 @@ class Daemon(object):
             Full path to the directory
 
         """
-        basedir = expand(
-            self.config.get('Directories', 'basedir').replace('<!User!>', usr)
-        )
+        basedir = self.config_parser.get('Directories', 'basedir')
+        basedir = expand(basedir.replace('<!User!>', usr))
         if dir_kind == 'basedir':
             return basedir
-        return join(basedir, self.config.get('Directories', dir_kind))
+        return join(basedir, self.config_parser.get('Directories', dir_kind))
 
     def mkdir(self, path):
         """Create a directory and set ownership and permissions appropriately.
@@ -334,9 +363,9 @@ class Daemon(object):
             else:
                 other += 1
 
-        self.queue_stat['q'] = queued
-        self.queue_stat['r'] = running
-        self.queue_stat['o'] = other
+        self.queue_status['q'] = queued
+        self.queue_status['r'] = running
+        self.queue_status['o'] = other
 
         # decide if queue is full
         if running < self.n_run and queued < self.n_queue:
@@ -361,7 +390,7 @@ class Daemon(object):
         if not jobs:
             wstderr('no jobs!\n')
             return
-        free = self.n_run - self.queue_stat['r']
+        free = self.n_run - self.queue_status['r']
 
         # Try couple of times to find some work, then go back to sleep (outer
         # loop)
@@ -493,6 +522,9 @@ class Daemon(object):
         pid_in_file = int(lines[-2].strip())
         host_in_file = lines[-1].strip()
 
+        if pid_in_file == self.pid:
+            return
+
         if host_in_file == self.hostname:
             wstderr('Attempting to kill process %d on host %s... '
                     % (pid_in_file, host_in_file))
@@ -510,6 +542,12 @@ class Daemon(object):
                 remove(self.pid_fpath)
                 return
 
+        #with open(expand('~/.openQ.err'), 'w') as fobj:
+        #    fobj.write('host_in_file = %s, self.hostname = %s\n'
+        #               % (host_in_file, self.hostname))
+        #set_path_metadata(expand('~/.openQ.err'), perms=0o660,
+        #                  group=self.group)
+
         raise OSError(
             'PID file already exists at "%s"! Kill possible existing'
             ' daemon with PID %d on host %s (attempt with kill before'
@@ -520,7 +558,9 @@ class Daemon(object):
 
     def set_my_umask_gid(self):
         """Set umask and--if one is defined--group ID (GID)"""
+        # Turn on user and group, disable "other" permissions
         umask(0)
+        # TODO: the following fails. Any way to setgid on process?
         if self.gid is not None:
             try:
                 setgid(self.gid)
@@ -541,10 +581,11 @@ class Daemon(object):
                 # Exit first parent
                 sys.exit(0)
         except OSError as err:
-            wstderr('fork #1 failed: %d (%s)\n' % (err.args[0], err.args[1]))
+            wstderr('Fork #1 failed: %d (%s)\n' % (err.args[0], err.args[1]))
             sys.exit(1)
 
         # Decouple from parent environment
+        chdir('/')
         setsid()
         self.set_my_umask_gid()
 
@@ -574,9 +615,14 @@ class Daemon(object):
         dup2(self.stderr.fileno(), sys.stderr.fileno())
 
         atexit.register(self.cleanup)
-        signal.signal(signal.SIGTERM, lambda signum, stack_frame: exit())
+        for sig in [signal.SIGTERM, signal.SIGHUP, signal.SIGINT,
+                    signal.SIGKILL]:
+            signal.signal(sig, sig_handler)
 
-        # Write PID and host to ~/.pid file
+        self.record_pid()
+
+    def record_pid(self):
+        """Write PID and host to ~/.pid file"""
         if self.pid_fpath is not None:
             with open(self.pid_fpath, 'w+') as pid_file:
                 pid_file.write('%s\n%s\n' % (self.pid, self.hostname))
@@ -584,26 +630,34 @@ class Daemon(object):
 
     def serve_forever(self):
         """Main loop"""
-        self.reconf()
+        self.kill_existing_daemon()
+        self.record_pid()
+
+        self.reconfigure()
         self.upgrade()
 
         if self.make_daemon:
             self.daemonize()
 
-        with open(expand('~/delmexyzabc'), 'w') as testfile:
-            testfile.write('\n')
+        # DEBUG
+        #with open(expand('~/PBS/delmexyzabc'), 'w') as testfile:
+        #    testfile.write('\n')
 
         while True:
-            if self.full:
-                wstderr('Queue is full.')
-            else:
-                self.do_some_work()
+            try:
+                if self.full:
+                    wstderr('Queue is full.')
+                else:
+                    self.do_some_work()
 
-            wstderr('going to sleep for %s seconds...\n' % self.sleep)
-            sleep(self.sleep)
+                wstderr('going to sleep for %s seconds...\n' % self.sleep)
+                sleep(self.sleep)
 
-            self.reconf()
-            self.upgrade()
+                self.reconfigure()
+                self.upgrade()
+            except:
+                self.cleanup()
+                raise
 
 
 def parse_args(description=__doc__):
@@ -617,6 +671,10 @@ def parse_args(description=__doc__):
         '--not-daemon', action='store_true',
         help='''Do *not* deamonize process (i.e., run interactively).'''
     )
+    parser.add_argument(
+        '--logfile', default=None, type=str,
+        help='''Write stdout and stderr to this logfile'''
+    )
     args = parser.parse_args()
     return args
 
@@ -624,7 +682,8 @@ def parse_args(description=__doc__):
 def main():
     """Main"""
     args = parse_args()
-    daemon = Daemon(configfile=args.config, daemon=(not args.not_daemon))
+    daemon = Daemon(configfile=args.config, daemon=(not args.not_daemon),
+                    logfile=args.logfile)
     daemon.serve_forever()
